requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorgram.py==1.2.0
debugpy==1.8.0
distlib==0.3.7
docopt==0.6.2
filelock==3.12.4
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==4.7.6
numpy==1.26.0
pandas==2.1.1
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.11.0
prettytable==3.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.5
wcwidth==0.2.8
Werkzeug==3.0.0
yarg==0.1.9
yarl==1.9.2
